requirements.txt
fastapi
uvicorn
sqlalchemy
openai
anthropic
google-generativeai
python-dotenv
jinja2

